+++ packages/jest-types/src/index.ts
@@ -8,5 +8,6 @@
 import type * as Circus from './Circus';
 import type * as Config from './Config';
 import type * as Global from './Global';
+import type * as TransformTypes from './Transform';
 
-export type {Circus, Config, Global};
+export type {Circus, Config, Global, TransformTypes};